+++ frontend/src/pages/GamePage/components/SelectContainer/GameVoteStatusContainer/GameVoteStatusContainer.styled.ts
