+++ frontend/src/pages/GamePage/components/SelectContainer/Timer/Timer.styled.ts
