+++ frontend/src/pages/GamePage/components/SelectContainer/Timer/Timer.util.ts
