+++ frontend/src/pages/GamePage/components/SelectContainer/Timer/hooks/useVoteTimer.ts
@@ -1,5 +1,5 @@
 import useTimer from './useTimer';
-import useCompleteSelectionMutation from '../../../components/SelectButton/SelectButton.hook';
+import useCompleteSelectionMutation from '../../SelectButton/SelectButton.hook';
 import { convertMsecToSecond } from '../Timer.util';
 
 import useBalanceContentQuery from '@/hooks/useBalanceContentQuery';
